// frontend/src/types/database.types.ts
export interface TaskDB {
  id: string;
  date_entered: Date | string | null;
  date_modified: Date | string | null;
  created_by: string | null;
  modified_user_id: string | null;
  assigned_user_id: string | null;
  deleted: number | null;
  name: string | null;
  status: string | null;
  priority: string | null;
  parent_type: string | null;
  parent_id: string | null;
  contact_id: string | null;
  date_due: Date | string | null;
  description: string | null;
}

export type TaskCreateRequest = Pick<TaskDB,
  'created_by' | 'modified_user_id' | 'assigned_user_id' | 'name' | 'status' | 'priority' | 'parent_type' | 'parent_id' | 'contact_id' | 'date_due' | 'description'
>;

export type TaskUpdateRequest = Partial<TaskCreateRequest>;

export type TaskResponse = TaskDB;

export interface CallDB {
  id: string;
  date_entered: Date | string | null;
  date_modified: Date | string | null;
  created_by: string | null;
  modified_user_id: string | null;
  assigned_user_id: string | null;
  deleted: number | null;
  name: string | null;
  duration_hours: number | null;
  duration_minutes: number | null;
  date_start: Date | string | null;
  status: string | null;
  direction: string | null;
  parent_type: string | null;
  parent_id: string | null;
  description: string | null;
}

export type CallCreateRequest = Pick<CallDB,
  'created_by' | 'modified_user_id' | 'assigned_user_id' | 'name' | 'duration_hours' | 'duration_minutes' | 'date_start' | 'status' | 'direction' | 'parent_type' | 'parent_id' | 'description'
>;

export type CallUpdateRequest = Partial<CallCreateRequest>;

export type CallResponse = CallDB;

export interface MeetingDB {
  id: string;
  date_entered: Date | string | null;
  date_modified: Date | string | null;
  created_by: string | null;
  modified_user_id: string | null;
  assigned_user_id: string | null;
  deleted: number | null;
  name: string | null;
  location: string | null;
  duration_hours: number | null;
  duration_minutes: number | null;
  date_start: Date | string | null;
  date_end: Date | string | null;
  status: string | null;
  parent_type: string | null;
  parent_id: string | null;
  description: string | null;
}

export type MeetingCreateRequest = Pick<MeetingDB,
  'created_by' | 'modified_user_id' | 'assigned_user_id' | 'name' | 'location' | 'duration_hours' | 'duration_minutes' | 'date_start' | 'date_end' | 'status' | 'parent_type' | 'parent_id' | 'description'
>;

export type MeetingUpdateRequest = Partial<MeetingCreateRequest>;

export type MeetingResponse = MeetingDB;

export interface NoteDB {
  id: string;
  date_entered: Date | string | null;
  date_modified: Date | string | null;
  created_by: string | null;
  modified_user_id: string | null;
  assigned_user_id: string | null;
  deleted: number | null;
  name: string | null;
  parent_type: string | null;
  parent_id: string | null;
  contact_id: string | null;
  description: string | null;
}